DevBuilding 0.0.1 Функция разделения интервала времени на значимые даты 2 версии в процессе написания 3я

+++ Assets/_Scripts/ChartScripts/Entities/CoordinateGrid.cs
@@ -33,7 +33,7 @@ namespace Chart
 
         public Action Updated;
 
-        public DateTime FromXAxisToDate(int x)
+        public DateTime FromXAxisToDate(float x)
         {
             return zeroPoint + x * step;
         }
