+++ Assets/_Scripts/ChartScripts/Interfaces/IGrid.cs
@@ -10,7 +10,7 @@ namespace Chart
 
         DateTime ZeroPoint { get; set; }
         TimeFrame Step { get; set; }
-        DateTime FromXAxisToDate(int x);
+        DateTime FromXAxisToDate(float x);
         int FromDateToXAxis(DateTime dateTime);
         float FromYAxisToPrice(float y);
         float FromPriceToYAxis(float price);
